split miner info into its own package
