+++ sia/miner/miner.go
@@ -1,7 +1,6 @@
 package miner
 
 import (
-	"encoding/json"
 	"runtime"
 	"sync"
 
@@ -27,49 +26,12 @@ type Miner struct {
 	sync.RWMutex
 }
 
-type Status struct {
-	State          string
-	Threads        int
-	RunningThreads int
-	Address        consensus.CoinAddress
-}
-
 // New takes a block channel down which it drops blocks that it mines. It also
 // takes a thread count, which it uses to spin up miners on separate threads.
 func New() (m *Miner) {
 	return &Miner{
 		iterationsPerAttempt: 256 * 1024,
 	}
-}
-
-// Info() returns a JSON struct which can be parsed by frontends for displaying
-// information to the user.
-func (m *Miner) Info() ([]byte, error) {
-	m.RLock()
-	defer m.RUnlock()
-
-	status := Status{
-		Threads:        m.threads,
-		RunningThreads: m.runningThreads,
-		Address:        m.address,
-	}
-
-	// Set the running status based on desiredThreads vs. runningThreads.
-	if m.desiredThreads == 0 && m.runningThreads == 0 {
-		status.State = "Off"
-	} else if m.desiredThreads == 0 && m.runningThreads > 0 {
-		status.State = "Turning Off"
-	} else if m.desiredThreads == m.runningThreads {
-		status.State = "On"
-	} else if m.desiredThreads < m.runningThreads {
-		status.State = "Turning On"
-	} else if m.desiredThreads > m.runningThreads {
-		status.State = "Decreasing number of threads."
-	} else {
-		status.State = "Miner is in an ERROR state!"
-	}
-
-	return json.Marshal(status)
 }
 
 // SubsidyAddress returns the address that is currently being used by the miner
